Fix: Move data providers
