+++ test/Unit/Format/FormatterTest.php
@@ -46,7 +46,7 @@ final class FormatterTest extends Framework\TestCase
     }
 
     /**
-     * @dataProvider \Ergebnis\Json\Normalizer\Test\Util\DataProvider\Boolean::provideBoolean()
+     * @dataProvider \Ergebnis\Json\Normalizer\Test\DataProvider\Boolean::provideBoolean()
      *
      * @param bool $hasFinalNewLine
      */
