+++ test/Unit/Validator/SchemaValidatorTest.php
@@ -35,7 +35,7 @@ final class SchemaValidatorTest extends Framework\TestCase
     }
 
     /**
-     * @dataProvider \Ergebnis\Json\Normalizer\Test\Util\DataProvider\Boolean::provideBoolean()
+     * @dataProvider \Ergebnis\Json\Normalizer\Test\DataProvider\Boolean::provideBoolean()
      *
      * @param bool $isValid
      */
